style: viewRoutes 분리 및 폴더 구조 수정
#7
