--- app/src/routes/home/index.js
@@ -1,10 +1,0 @@
-"use strict";
-
-const express = require("express");
-const router = express.Router();
-
-const ctrl = require("../../controllers/home");
-
-router.get("/", ctrl.output.home);
-
-module.exports = router;
